Adding Deconvolution layer.
Currently target_shape has been compulsory since it is required by coreml. Also, we are not currently able to evaluate the affect of padding. If we try to xplicitly add padding to coreml model at the end we get an error. This is how we were adding the padding:
         pad = literal_eval(param['pad'])
         for i in range(len(pad)):
             convLayer.valid.paddingAmounts.borderAmounts[i].startEdgeSize = pad[i]
             convLayer.valid.paddingAmounts.borderAmounts[i].endEdgeSize = pad[i]

Error:
  File "test_mxnet_converer.py", line 22, in _get_coreml_model
    spec = mxnet_converter.convert(model, class_labels=class_labels, mode=mode, **input_shape)
  File "_mxnet_converter.py", line 197, in convert
    converter_func(net, node, model, builder)
  File "_layers.py", line 515, in convert_deconvolution
    convLayer.valid.paddingAmounts.borderAmounts[i].startEdgeSize = pad[i]
  File "/Users/pracheer/miniconda3/envs/coremltools/lib/python2.7/site-packages/google/protobuf/internal/containers.py", line 204, in __getitem__
    return self._values[key]
IndexError: list index out of range

Will fix the above two issues after discussions with coreml people.